Add support for rowing activities
To a first approximation, this is just about adding it to the filter;
all the activities are already downloaded, so we just didn't show them
before. Rowing (in particular indoor rowing) is useful for runners that
use it as cross-training or during injury recovery.

However, rowing pace is traditionally given in time per 500m, not time
per kilometer as running typically is—even though it's roughly in
the same range. (E.g., I usually row somewhere around 1:55/500m,
which is more intense for me than running at 3:50/km, but both are
entirely doable.) Thus, we make an effort to switch pace; the pace
formatters (both in C# and in TypeScript) now support taking in an
activity type, and when it's rowing, we switch to pace/500m.

When we show values pertinent to multiple activities (e.g. average
pace per week), we use the more universal pace/km, but when we
show individual activities, or we filter only on rowing activities,
we switch to pace/500m even for averages and such. Graphed values
are of course still based on km/h, so the bars don't get larger or
smaller based on this.