Try to merge autolaps.
Before doing interval detection or Bislett rounds, try to detect autolaps
(multiples of a whole kilometer, or nearly so) and merge them into the
neighboring lap. This makes the code much more precise in practice;
I've run it over all of my activities for the last four years, and there's
a clear net gain. (E.g. a run of 3x3000m with autolap on was previously
not even detected as an interval run, whereas after merging, it shows up
and has merged the autolaps correctly.)

Of course, there are still situations where warm-ups show incorrectly
as intervals, and there are activities where this has a net negative
effect (e.g., a 1K+1K+5K+1K run used to show up as eight 1K intervals,
but after merging the 5K, we no longer think it qualifies as an interval
and show only the three 1Ks). Of course, for users that never ever use
autolap nor lap manually at 1K intervals, it will probably never do any
good. In general, however, it seems to be a clear win at least on
my dataset.