Support manual overrides of interval speed.
This allows people to write (in either the activity description or
in private notes, as usual) some special text to override the speed
of an interval or a range of intervals. The intended use is for
users who run on treadmill without good footpod measurement.
One can use either pace or speed notation, in Norwegian or English.

Distance is adjusted accordingly, and global activity values are
recalculated from the new laps. All of this happens before any other
processing, so that e.g. interval lap detection gets to work on the
(presumably) correct data.